added gui

+++ common/enc.cpp
@@ -1,7 +1,7 @@
 #include "enc.h"
 
 Encryption::Encryption(string text) {
-    for(int i =0 ; i < text.size() ; i++) {
+    for(int i =0 ; i < static_cast<int>(text.size()) ; i++) {
         text[i] += 2;
     }
     encPass = text;
